requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
pydantic>=2.8.0
PyMuPDF>=1.24.0
python-dotenv>=1.0.1
requests>=2.32.0
pandas>=2.2.2
